Fixed the seconds not displaying by code introduced to fix a crash on a simulated FR55 when in low power mode and seconds should be displayed
The real cause of the error is in the 208x208 resource file here

		<!-- Partial updates not supported -->
		<!--param name="secondsClipY">139</param-->
		<!--param name="secondsClipWidth">22</param-->
		<!--param name="secondsClipHeight">15</param-->

With the variables commented out and the FR55 supporting 'partial update', it crashes. No one complained about the frozen seconds in low power mode so I'm leaving it like that for the time being. Might fix later.