"Hide seconds" has been changed from a checkbox to a list with three entries
  - Show seconds
  - Hide seconds when inactive
  - Always hide seconds

Show and Hide seconds behaves like the old toggle but "Hide seconds when inactive" will not display seconds when the watchface is in low power mode (ie, inactive). "Memory-In-Pixel" watch consumes energy when a pixel is changed. Preventing the seconds to update every secondes can help reduce the energy drawn by the watchface. Has no Effect on AMOLED watches.